Add web-assume-role
This provides a federated login url to the aws web console.

Split sessionTimeout and roletimeout. These were stomping on each other
in a surprising ways. The last *Command() handler was setting the
final value on the timeout which meant we were setting a 30 minute
timeout for a role login. This was surprising.