Support other aws partitions
Add support for the other aws partitions (gov, china, etc.). This is
configured on the profile in the config.

Rename `provider` to `profile`. I was originally thinking about this
in terms of secrets providers (op,pass,etc.). Now I'm thinking about
this in terms of base accounts. You have a single set of long term
credentials stored in a password manager for an account.

Also --provider_id was too hard to remember.